refactor: move component to correct folder
