+++ apps/ui-kit/src/lib/components/atoms/button-segment/ButtonSegment.tsx
@@ -12,7 +12,7 @@ import {
 } from './button-segment.classes';
 import cx from 'classnames';
 import { ButtonSegmentType } from './button-segment.enums';
-import { UnstyledButton } from '../button/Button';
+import { ButtonUnstyled } from '../button/ButtonUnstyled';
 
 interface ButtonSegmentProps {
     /**
@@ -63,7 +63,7 @@ export function ButtonSegment({
     const borderRadius = isUnderlined ? '' : 'rounded-full';
     const textSize = isUnderlined ? 'text-title-md' : 'text-label-lg';
     return (
-        <UnstyledButton
+        <ButtonUnstyled
             onClick={onClick}
             className={cx(
                 'enabled:state-layer relative flex items-center disabled:opacity-40',
@@ -82,6 +82,6 @@ export function ButtonSegment({
                 {icon && <span>{icon}</span>}
                 <span className="font-inter">{label}</span>
             </div>
-        </UnstyledButton>
+        </ButtonUnstyled>
     );
 }
