+++ apps/ui-kit/src/lib/components/atoms/index.ts
@@ -1,14 +1,12 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './address';
 export * from './badge';
 export * from './button';
 export * from './radio-button';
 export * from './button-segment';
 export * from './button';
 export * from './checkbox';
-export * from './chip';
 export * from './divider';
 export * from './key-value-info';
 export * from './label-text';
