+++ apps/ui-kit/src/lib/components/molecules/account/Account.tsx
@@ -4,8 +4,9 @@
 import React from 'react';
 import cx from 'classnames';
 import { Button, ButtonSize, ButtonType } from '../../atoms/button';
-import { Address, Badge, BadgeType } from '../../atoms';
+import { Badge, BadgeType } from '../../atoms';
 import { LockLocked, LockUnlocked, MoreHoriz } from '@iota/ui-icons';
+import { Address } from '../address';
 
 interface AccountProps {
     /**
@@ -39,11 +40,11 @@ interface AccountProps {
     /**
      * The onCopy event of the Address  (optional).
      */
-    onCopy?: (e: React.MouseEvent<SVGElement>) => void;
+    onCopy?: (e: React.MouseEvent<HTMLButtonElement>) => void;
     /**
      * The onOpen event of the Address  (optional).
      */
-    onOpen?: (e: React.MouseEvent<SVGElement>) => void;
+    onOpen?: (e: React.MouseEvent<HTMLButtonElement>) => void;
     /**
      * Has copy icon (optional).
      */
