+++ apps/ui-kit/src/lib/components/molecules/index.ts
@@ -1,6 +1,8 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+export * from './address';
+export * from './chip';
 export * from './segmented-button';
 export * from './title';
 export * from './input';
